Moved the QueryString class to the Tool directory
